requirements.txt
selenium>=4.10.0
pyperclip>=1.8.2
Pillow>=10.0.0
